make table ID being strongly typed

+++ pkg/sqlstore/impl/system/migrations/005_mul_changes.up.sql
@@ -1,6 +1,6 @@
 ALTER TABLE system_tables DROP COLUMN type;
 ALTER TABLE system_tables DROP COLUMN uuid;
-ALTER TABLE system_tables ADD COLUMN id BIGINT PRIMARY KEY CHECK (id >= 0);
+ALTER TABLE system_tables ADD COLUMN id NUMERIC(60) PRIMARY KEY CHECK (id >= 0);
 ALTER TABLE system_tables ADD COLUMN structure CHAR(64) NOT NULL;
 ALTER TABLE system_tables ADD COLUMN description VARCHAR(100) NOT NULL;
 ALTER TABLE system_tables ADD COLUMN name VARCHAR(50) NOT NULL;
